travis: Don't lie about which Git SHA is being tagged.
It is unsafe to limit the evaluated history to only one subdirectory,
as this will cause the container to be rebuilt with the same SHA,
even if there is a new SHA in the history. The intention of limiting
to the service directory was presumably to avoid unnecessary rebuilds,
but this is dangerous, since the resulting containers are used in
Hosted Mender, and we cannot guarantee that Docker produces exactly
the same image. In fact, in most cases it won't.

changelog: none

Signed-off-by: Krzysztof Jaskiewicz <krzysiekjaskiewicz@gmail.com>